Move browser to vscodeApis
